Reworked config system, parents and interfaces
