+++ docs/JavaDoc/net/vortexdata/tsqpf/configs/ConfigField.html
@@ -94,7 +94,7 @@ try {
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>net.vortexdata.tsqpf.configs.ConfigField</li>
+                <li>net.vortexdata.tsqpf.configs.Config</li>
             </ul>
         </li>
     </ul>
