+++ docs/JavaDoc/net/vortexdata/tsqpf/configs/ConfigMain.html
@@ -100,7 +100,7 @@ var activeTableTab = "activeTableTab";
         <li>
             <ul class="inheritance">
                 <li><a href="../../../../net/vortexdata/tsqpf/configs/ConfigField.html"
-                       title="class in net.vortexdata.tsqpf.configs">net.vortexdata.tsqpf.configs.ConfigField</a></li>
+                       title="class in net.vortexdata.tsqpf.configs">net.vortexdata.tsqpf.configs.Config</a></li>
                 <li>
                     <ul class="inheritance">
                         <li>net.vortexdata.tsqpf.configs.ConfigMain</li>
