+++ src/main/java/net/vortexdata/tsqpf/configs/Config.java
@@ -1,22 +1,70 @@
 package net.vortexdata.tsqpf.configs;
 
 
+import java.io.File;
+import java.io.FileInputStream;
 import java.util.HashMap;
+import java.util.Properties;
+import java.util.Set;
 
 /**
- * Configuration parent class prototype
+ * Config fields prototype
  *
  * @author Sandro Kierner
- * @since 1.0.0
+ * @since 2.0.0
  */
-public interface Config {
+public class Config implements ConfigInterface {
 
-    boolean load();
+    protected String path = "";
+    protected HashMap<String, String> values;
+    protected HashMap<String, String> defaultValues;
 
-    HashMap<String, String> getValues();
+    public Config(String path) {
+        this.path = path;
+    }
 
-    HashMap<String, String> getDefaultValues();
+    public boolean load() {
 
-    String getPath();
+        HashMap<String, String> values = new HashMap<String, String>();
+        File file = new File(path);
+        // net.vortexdata.tsqpf.Test if config exists
+        if (!file.exists()) {
+            // Return empty map if config is empty.
+            values = null;
+            return false;
+        } else {
+            try {
+                Properties prop = new Properties();
+                prop.load(new FileInputStream(path));
+
+                // Load all keys & values
+                Set<Object> keys = prop.keySet();
+                for (Object k : keys) {
+                    values.put((String) k, prop.getProperty((String) k));
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+
+        this.values = values;
+        return true;
+
+    }
+
+    @Override
+    public HashMap<String, String> getValues() {
+        return values;
+    }
+
+    @Override
+    public HashMap<String, String> getDefaultValues() {
+        return defaultValues;
+    }
+
+    @Override
+    public String getPath() {
+        return path;
+    }
 
 }
