+++ src/main/java/net/vortexdata/tsqpf/configs/ConfigInterface.java
@@ -1,0 +1,22 @@
+package net.vortexdata.tsqpf.configs;
+
+
+import java.util.HashMap;
+
+/**
+ * Configuration parent class prototype
+ *
+ * @author Sandro Kierner
+ * @since 1.0.0
+ */
+public interface ConfigInterface {
+
+    boolean load();
+
+    HashMap<String, String> getValues();
+
+    HashMap<String, String> getDefaultValues();
+
+    String getPath();
+
+}
