+++ src/main/java/net/vortexdata/tsqpf/configs/ConfigMain.java
@@ -1,7 +1,5 @@
 package net.vortexdata.tsqpf.configs;
 
-import net.vortexdata.tsqpf.installers.InstallConfig;
-
 import java.util.HashMap;
 
 /**
@@ -10,8 +8,7 @@ import java.util.HashMap;
  * @author Sandro Kierner
  * @since 1.0.0
  */
-public class ConfigMain extends ConfigField implements Config {
-
+public class ConfigMain extends Config {
 
     public ConfigMain() {
         super("configs//main.properties");
@@ -27,44 +24,7 @@ public class ConfigMain extends ConfigField implements Config {
         defaultValues.put("clientNickname", "TSQP Framework");
         defaultValues.put("reconnectStrategy", "exponentialBackoff");
         defaultValues.put("messageChatCommandNotFound", "Unknown command.");
-    }
-
-    /**
-     * Loads the configs file.
-     *
-     * @return true if default value count and loaded value count are equal
-     */
-    public boolean load() {
-        ConfigLoader cLoader = new ConfigLoader();
-        values = cLoader.load(path);
-        if (values.size() < defaultValues.size()) {
-            InstallConfig iConfig = new InstallConfig();
-            iConfig.create(path, defaultValues);
-            return false;
-        }
-        return true;
-    }
-
-    /**
-     * Returns a config value
-     *
-     * @param key Key of value
-     * @return Desired value of key
-     */
-    public String getProperty(String key) {
-        return values.get(key);
-    }
-
-    public HashMap<String, String> getValues() {
-        return null;
-    }
-
-    public HashMap<String, String> getDefaultValues() {
-        return null;
-    }
-
-    public String getPath() {
-        return null;
+        defaultValues.put("shellPort", "12342");
     }
 
 }
